+++ packages/tooltip/index.js
@@ -1,1 +1,2 @@
 export { LionTooltip } from './src/LionTooltip.js';
+export { LionTooltipArrow } from './src/LionTooltipArrow.js';