+++ packages/form-system/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.14](https://github.com/ing-bank/lion/compare/@lion/form-system@0.2.13...@lion/form-system@0.2.14) (2019-12-16)
+
+**Note:** Version bump only for package @lion/form-system
+
+
+
+
+
 ## [0.2.13](https://github.com/ing-bank/lion/compare/@lion/form-system@0.2.12...@lion/form-system@0.2.13) (2019-12-13)
 
 **Note:** Version bump only for package @lion/form-system