+++ packages/overlays/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.10.1](https://github.com/ing-bank/lion/compare/@lion/overlays@0.10.0...@lion/overlays@0.10.1) (2019-12-17)
+
+**Note:** Version bump only for package @lion/overlays
+
+
+
+
+
 # [0.10.0](https://github.com/ing-bank/lion/compare/@lion/overlays@0.9.0...@lion/overlays@0.10.0) (2019-12-13)